condense s3 events
