+++ shared/util.js
@@ -69,9 +69,48 @@ var isRequestValid = function(options) {
   return true;
 };
 
+/**
+ * Extends a destination object with the properties of one or more source
+ * objects.
+ * @param {Object} destination The destination object.
+ * @param {...Object} var_sources One or more source objects.
+ * @return {Object} The original destination object, with all the additional
+ *     properties from the source objects.
+ */
+var extend = function(destination, var_sources) {
+  var sources = Array.prototype.slice.call(arguments, 1);
+
+  sources.forEach(function(source) {
+    if (!source) { return; }
+
+    for (var property in source) {
+      destination[property] = source[property];
+    }
+  });
+
+  return destination;
+};
+
+/**
+ * Creates a metadata object representing a directory.
+ * @param {string} name The name of the directory.
+ * @return {Object} A plain object representing a directory with the given name
+ *     in the file system provider API.
+ */
+var makeDirectory = function(name) {
+  return {
+    isDirectory: true,
+    name: name,
+    size: 0,
+    modificationTime: new Date(0),
+  };
+};
+
 module.exports = {
   arrayBufferToString: arrayBufferToString,
   stringToArrayBuffer: stringToArrayBuffer,
   blobToArrayBuffer: blobToArrayBuffer,
-  isRequestValid: isRequestValid
+  isRequestValid: isRequestValid,
+  extend: extend,
+  makeDirectory: makeDirectory
 };
